removed universal styledComponents infavour of page specific styledComponets
