+++ src/Components/Shared/styles/Dogs.styled.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+import { Row } from 'react-bootstrap';
+
+export const StyledRow = styled(Row)`
+  align-items: center;
+  justify-content: center;
+`;
