+++ src/Components/Shared/styles/Header.styled.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+import { Container, Image } from 'react-bootstrap';
+
+// Header Custom Styling
+export const HeaderContainer = styled(Container)`
+  display: flex;
+  justify-content: center;
+  padding: 0;
+`;
+
+export const HeaderImage = styled(Image)`
+max-width: 100%;
+max-height: 100%;
+`;
